Shift concept activations to model forward pass
